removed 'identifier' from blake2b function, replaced with nn for more flexible digest length. Removed the memory clearing, will do it in argon2 function.

+++ blake2b.h
@@ -4,7 +4,7 @@
 #include <inttypes.h>
 #include <stddef.h>
 
-uint8_t *blake2b(uint8_t*, size_t, uint16_t);
+uint8_t *blake2b(uint8_t*, size_t, uint8_t);
 
 #endif
